refactor(event): recurrence_service.py 578行を recurrence/ パッケージに分割
単一ファイルに集中していた RecurrenceService の責務を機能別に分離し、
将来の差分レビューと変更の局所化を容易にする。

分割後:
- recurrence/parser.py (126): カスタムルール文字列の解析と deterministic spec 抽出
- recurrence/calculator.py (221): 日付計算（rule ベース / spec ベース / 第N曜日）
- recurrence/llm_generator.py (193): LLM 経由の日付生成と直近開催履歴フォーマット
- recurrence/persistence.py (69): Event テーブルへのマスター + インスタンス作成
- recurrence/service.py (178): RecurrenceService 本体（上記を組み合わせるファサード）
- recurrence_service.py (29): 互換シム（Event 含めて re-export）

旧 import パス (event.recurrence_service.RecurrenceService) と
旧 patch ターゲット (event.recurrence_service.Event.objects.filter,
event.recurrence_service.RecurrenceService.generate_dates) は維持。
event.recurrence_service 名のロガーも llm_generator.py 内で明示再現。

検証: event アプリ全 377 件 pass。
