+++ app/event/recurrence_service.py
@@ -1,578 +1,29 @@
-"""定期イベントの日付生成サービス"""
+"""recurrence_service 互換シム (後方互換性のため re-export)
+
+新規コードは `from event.recurrence import RecurrenceService` を使うこと。
+
+ここで `Event` も re-export しているのは、既存テストが
+`patch("event.recurrence_service.Event.objects.filter", ...)` の形で
+モックしているため。`event.models.Event` と同一オブジェクトを指すので、
+実装が `event/recurrence/llm_generator.py` 側にあっても patch は効く。
+
+`event.recurrence_service` 名のロガーも維持。実際の出力は
+`llm_generator.py` 内の `logging.getLogger("event.recurrence_service")` から行う。
+"""
 import logging
-import re
-from calendar import monthrange
-from datetime import datetime, timedelta, date
-from typing import TYPE_CHECKING, Dict, List, Optional
 
-from django.db import models
-
-from event.llm_service import EventDateLlmService, get_event_date_llm_service
-from event.models import Event, RecurrenceRule
-
-if TYPE_CHECKING:
-    from community.models import Community
+from event.models import Event  # noqa: F401  (patch 用に re-export)
+from event.recurrence import RecurrenceService  # noqa: F401
+from event.recurrence.calculator import (  # noqa: F401  (旧 import 互換)
+    get_japanese_weekday,
+    get_nth_weekday_of_month,
+    get_week_of_month,
+)
+from event.recurrence.parser import (  # noqa: F401  (旧 import 互換)
+    WEEK_TOKEN_MAP,
+    WEEKDAY_TOKEN_MAP,
+)
 
 logger = logging.getLogger(__name__)
 
-WEEKDAY_TOKEN_MAP = {
-    '月': 0,
-    '火': 1,
-    '水': 2,
-    '木': 3,
-    '金': 4,
-    '土': 5,
-    '日': 6,
-}
-
-WEEK_TOKEN_MAP = {
-    '1': 1,
-    '2': 2,
-    '3': 3,
-    '4': 4,
-    '5': 5,
-    '一': 1,
-    '二': 2,
-    '三': 3,
-    '四': 4,
-    '五': 5,
-    '最終': -1,
-}
-
-
-class RecurrenceService:
-    """定期イベントの日付を生成するサービス"""
-
-    def __init__(self, llm_service: Optional[EventDateLlmService] = None):
-        self._llm_service = llm_service
-
-    def _get_llm_service(self) -> EventDateLlmService:
-        if self._llm_service is None:
-            self._llm_service = get_event_date_llm_service()
-        return self._llm_service
-    
-    def generate_dates(self, rule: RecurrenceRule, base_date: date, base_time: datetime.time, months: int = 1, community: Optional["Community"] = None) -> List[date]:
-        """定期ルールに基づいて日付リストを生成"""
-        if rule.frequency in ['WEEKLY', 'MONTHLY_BY_DATE', 'MONTHLY_BY_WEEK']:
-            # ルールベースで生成
-            return self._generate_dates_by_rule(rule, base_date, months)
-        elif rule.frequency == 'OTHER':
-            deterministic_spec = self._parse_deterministic_custom_rule(rule)
-            if deterministic_spec is not None:
-                return self._generate_dates_by_deterministic_spec(
-                    deterministic_spec=deterministic_spec,
-                    base_date=base_date,
-                    months=months,
-                    end_date=rule.end_date,
-                )
-            # LLMで生成
-            return self._generate_dates_by_llm(rule, base_date, base_time, months, community)
-        return []
-
-    def has_deterministic_custom_rule(self, rule: RecurrenceRule) -> bool:
-        """サーバー側で安全に解釈できるカスタムルールかどうかを返す"""
-        return self._parse_deterministic_custom_rule(rule) is not None
-
-    def matches_custom_rule_date(self, rule: RecurrenceRule, check_date: date) -> bool:
-        """カスタムルールに一致する日付かを返す"""
-        deterministic_spec = self._parse_deterministic_custom_rule(rule)
-        if deterministic_spec is None:
-            return True
-        return self._matches_deterministic_spec(deterministic_spec, check_date)
-
-    def _normalize_custom_rule(self, custom_rule: Optional[str]) -> str:
-        if not custom_rule:
-            return ''
-
-        normalized_rule = custom_rule.translate(str.maketrans('０１２３４５６７８９', '0123456789'))
-        return re.sub(r'\s+', '', normalized_rule)
-
-    def _parse_deterministic_custom_rule(self, rule: RecurrenceRule) -> Optional[Dict]:
-        """既知の自由記述ルールを deterministic に解釈する"""
-        normalized_rule = self._normalize_custom_rule(rule.custom_rule)
-        if not normalized_rule:
-            return None
-
-        monthly_dates = self._extract_monthly_dates(normalized_rule)
-        if monthly_dates:
-            return {
-                'kind': 'monthly_dates',
-                'days': monthly_dates,
-            }
-
-        monthly_weekdays = self._extract_monthly_weekday_rules(normalized_rule)
-        if monthly_weekdays:
-            return {
-                'kind': 'monthly_weekdays',
-                'pairs': monthly_weekdays,
-            }
-
-        if normalized_rule in {'毎月', '月1'} and rule.start_date:
-            return {
-                'kind': 'monthly_weekdays',
-                'pairs': [(self._get_week_of_month(rule.start_date), rule.start_date.weekday())],
-            }
-
-        return None
-
-    def _extract_monthly_dates(self, normalized_rule: str) -> List[int]:
-        """日付指定の自由記述ルールを抽出する"""
-        single_day_match = re.fullmatch(r'毎月(\d{1,2})日(?:開催)?', normalized_rule)
-        if single_day_match:
-            return [int(single_day_match.group(1))]
-
-        date_tokens = [
-            int(token)
-            for token in re.findall(r'(?<!第)(\d{1,2})日', normalized_rule)
-        ]
-        valid_dates = sorted({token for token in date_tokens if 1 <= token <= 31})
-        if len(valid_dates) >= 2:
-            return valid_dates
-        return []
-
-    def _extract_monthly_weekday_rules(self, normalized_rule: str) -> List[tuple[int, int]]:
-        """第N曜日系の自由記述ルールを抽出する"""
-        pairs: list[tuple[int, int]] = []
-
-        for segment in re.split(r'[、,]', normalized_rule):
-            cleaned_segment = segment.split('：')[-1]
-            cleaned_segment = cleaned_segment.strip('（）()')
-
-            last_match = re.search(r'最終([月火水木金土日])曜日', cleaned_segment)
-            if last_match:
-                pairs.append((-1, WEEKDAY_TOKEN_MAP[last_match.group(1)]))
-                continue
-
-            week_day_match = re.search(
-                r'(第?(?:[1-5一二三四五]|最終)(?:・第?(?:[1-5一二三四五]|最終))*)([月火水木金土日])曜日',
-                cleaned_segment,
-            )
-            if week_day_match:
-                weekday = WEEKDAY_TOKEN_MAP[week_day_match.group(2)]
-                for week_token in week_day_match.group(1).split('・'):
-                    normalized_week = week_token.replace('第', '')
-                    week = WEEK_TOKEN_MAP.get(normalized_week)
-                    if week is not None:
-                        pairs.append((week, weekday))
-
-        return sorted(set(pairs))
-
-    def _generate_dates_by_deterministic_spec(
-        self,
-        deterministic_spec: Dict,
-        base_date: date,
-        months: int,
-        end_date: Optional[date] = None,
-    ) -> List[date]:
-        """deterministic に解釈できる自由記述ルールの日付を生成する"""
-        generated_dates = []
-        generation_end = base_date + timedelta(days=months * 30)
-        if end_date and end_date < generation_end:
-            generation_end = end_date
-
-        year = base_date.year
-        month = base_date.month
-        while True:
-            current_month_dates = self._resolve_month_dates(
-                deterministic_spec=deterministic_spec,
-                year=year,
-                month=month,
-            )
-
-            if current_month_dates and min(current_month_dates) > generation_end:
-                break
-
-            for current_date in current_month_dates:
-                if base_date <= current_date <= generation_end:
-                    generated_dates.append(current_date)
-
-            month += 1
-            if month > 12:
-                month = 1
-                year += 1
-
-            if date(year, month, 1) > generation_end:
-                break
-
-        return sorted(set(generated_dates))
-
-    def _resolve_month_dates(self, deterministic_spec: Dict, year: int, month: int) -> List[date]:
-        """指定月の開催日一覧を返す"""
-        if deterministic_spec['kind'] == 'monthly_dates':
-            last_day = monthrange(year, month)[1]
-            days = deterministic_spec['days']
-            if len(days) == 1:
-                return [date(year, month, min(days[0], last_day))]
-            return [date(year, month, day) for day in days if day <= last_day]
-
-        if deterministic_spec['kind'] == 'monthly_weekdays':
-            month_start = date(year, month, 1)
-            resolved_dates = []
-            for week, weekday in deterministic_spec['pairs']:
-                resolved_date = self._get_nth_weekday_of_month(month_start, weekday, week)
-                if resolved_date:
-                    resolved_dates.append(resolved_date)
-            return sorted(set(resolved_dates))
-
-        return []
-
-    def _matches_deterministic_spec(self, deterministic_spec: Dict, check_date: date) -> bool:
-        """deterministic spec に一致するか判定する"""
-        return check_date in self._resolve_month_dates(
-            deterministic_spec=deterministic_spec,
-            year=check_date.year,
-            month=check_date.month,
-        )
-    
-    def _generate_dates_by_rule(self, rule: RecurrenceRule, base_date: date, months: int) -> List[date]:
-        """ルールベースで日付リストを生成"""
-        dates = []
-        end_date = base_date + timedelta(days=months * 30)
-        
-        if rule.end_date and rule.end_date < end_date:
-            end_date = rule.end_date
-        
-        current_date = base_date
-        
-        if rule.frequency == 'WEEKLY':
-            # 毎週または隔週
-            if rule.start_date:
-                # start_dateの曜日を基準に、base_date以降の最初の同じ曜日を探す
-                target_weekday = rule.start_date.weekday()
-                current_date = base_date
-                
-                # base_dateから最も近い同じ曜日を探す
-                days_ahead = (target_weekday - current_date.weekday()) % 7
-                if days_ahead == 0 and current_date < base_date:
-                    days_ahead = 7
-                current_date = current_date + timedelta(days=days_ahead)
-                
-                # その日が開催日でない場合は、次の開催日まで進める
-                while not rule.is_occurrence_date(current_date):
-                    current_date += timedelta(weeks=1)
-            else:
-                current_date = base_date
-            
-            # 開催日を収集
-            while current_date <= end_date:
-                if rule.is_occurrence_date(current_date):
-                    dates.append(current_date)
-                current_date += timedelta(weeks=1)  # 1週間ずつ進めて、is_occurrence_dateで判定
-        
-        elif rule.frequency == 'MONTHLY_BY_DATE':
-            # 毎月（日付指定）
-            while current_date <= end_date:
-                dates.append(current_date)
-                # 次の月の同じ日付へ
-                if current_date.month == 12:
-                    next_month = 1
-                    next_year = current_date.year + 1
-                else:
-                    next_month = current_date.month + rule.interval
-                    next_year = current_date.year
-                
-                try:
-                    current_date = current_date.replace(year=next_year, month=next_month)
-                except ValueError:
-                    # 月末の場合（例：1月31日→2月28日）
-                    current_date = current_date.replace(year=next_year, month=next_month, day=1)
-                    current_date = (current_date + timedelta(days=32)).replace(day=1) - timedelta(days=1)
-        
-        elif rule.frequency == 'MONTHLY_BY_WEEK':
-            # 毎月（第N曜日）
-            # 使用する曜日を決定（start_dateがあればその曜日、なければ基準日の曜日を使用）
-            if rule.start_date:
-                target_weekday = rule.start_date.weekday()
-            else:
-                target_weekday = current_date.weekday()
-            
-            # 最初の日付を第N曜日に調整
-            first_date = self._get_nth_weekday_of_month(
-                current_date.replace(day=1),
-                target_weekday,
-                rule.week_of_month or 1
-            )
-            if first_date and first_date >= base_date:
-                current_date = first_date
-            else:
-                # 次の月から開始
-                next_month = (current_date.month % 12) + 1
-                next_year = current_date.year + (1 if next_month == 1 else 0)
-                current_date = self._get_nth_weekday_of_month(
-                    date(next_year, next_month, 1),
-                    target_weekday,
-                    rule.week_of_month or 1
-                )
-            
-            while current_date and current_date <= end_date:
-                dates.append(current_date)
-                # 次の月の第N曜日を計算
-                next_month = (current_date.month % 12) + 1
-                next_year = current_date.year + (1 if next_month == 1 else 0)
-                next_date = self._get_nth_weekday_of_month(
-                    date(next_year, next_month, 1),
-                    target_weekday,
-                    rule.week_of_month or 1
-                )
-                if next_date:
-                    current_date = next_date
-                else:
-                    break
-        
-        return dates
-    
-    def _get_japanese_weekday(self, weekday: int) -> str:
-        """曜日番号を日本語の曜日に変換"""
-        weekdays = ['月曜日', '火曜日', '水曜日', '木曜日', '金曜日', '土曜日', '日曜日']
-        return weekdays[weekday]
-    
-    def _get_nth_weekday_of_month(self, target_date: date, weekday: int, n: int) -> Optional[date]:
-        """指定月の第N曜日を取得"""
-        if n == -1:
-            last_day = target_date.replace(day=monthrange(target_date.year, target_date.month)[1])
-            days_back = (last_day.weekday() - weekday) % 7
-            return last_day - timedelta(days=days_back)
-
-        # 月初の日付を取得
-        first_day = target_date.replace(day=1)
-        
-        # 月初から最初の指定曜日までの日数
-        days_until_weekday = (weekday - first_day.weekday()) % 7
-        
-        # 第N曜日の日付を計算
-        nth_weekday = first_day + timedelta(days=days_until_weekday + (n - 1) * 7)
-        
-        # 同じ月内かチェック
-        if nth_weekday.month == target_date.month:
-            return nth_weekday
-        return None
-    
-    def _get_week_of_month(self, date_obj: date) -> int:
-        """日付が月の第何週かを取得（その曜日の第何回目の出現か）"""
-        # 同じ曜日の第何回目かを計算
-        # 例: 2024-11-25（月曜日）は11月の第4月曜日
-        count = 0
-        for day in range(1, date_obj.day + 1):
-            check_date = date_obj.replace(day=day)
-            if check_date.weekday() == date_obj.weekday():
-                count += 1
-        return count
-    
-    def _get_recent_events_history(self, rule: RecurrenceRule, base_date: date, community: Optional["Community"] = None) -> str:
-        """直近5回の開催履歴を取得してフォーマット"""
-        try:
-            recent_events = []
-            
-            # このルールに関連するイベントを取得
-            if rule.id:  # 既存のルールの場合
-                # RecurrenceRuleに紐づくマスターイベントを探す
-                master_events = Event.objects.filter(recurrence_rule=rule, is_recurring_master=True)
-                
-                if master_events.exists():
-                    # マスターイベントに紐づく過去のイベントを取得
-                    for master_event in master_events:
-                        # マスターイベント自身と、そのインスタンスを取得
-                        events = Event.objects.filter(
-                            models.Q(id=master_event.id) | models.Q(recurring_master=master_event),
-                            date__lt=base_date
-                        ).order_by('-date')[:5]
-                        recent_events.extend(events)
-                    
-                    # 全体でソートして最新5件を取得
-                    recent_events = sorted(recent_events, key=lambda e: e.date, reverse=True)[:5]
-            
-            # ルールに紐づくイベントがない場合、またはcommunityが指定されている場合
-            if not recent_events and community:
-                # コミュニティの過去のイベントから取得
-                recent_events = Event.objects.filter(
-                    community=community,
-                    date__lt=base_date
-                ).order_by('-date')[:5]
-            
-            if not recent_events:
-                return "過去の開催履歴: なし"
-            
-            # 履歴をフォーマット
-            history_lines = ["過去の開催履歴（直近5回）:"]
-            for event in reversed(recent_events):  # 古い順に表示
-                weekday = self._get_japanese_weekday(event.date.weekday())
-                week = self._get_week_of_month(event.date)
-                history_lines.append(
-                    f"- {event.date.strftime('%Y-%m-%d')} ({weekday}) 第{week}週"
-                )
-            
-            # パターン分析を追加
-            if len(recent_events) >= 2:
-                history_lines.append("\n開催パターン分析:")
-                
-                # 曜日の傾向
-                weekdays = [e.date.weekday() for e in recent_events]
-                weekday_counts = {}
-                for wd in weekdays:
-                    wd_name = self._get_japanese_weekday(wd)
-                    weekday_counts[wd_name] = weekday_counts.get(wd_name, 0) + 1
-                
-                if weekday_counts:
-                    most_common_weekday = max(weekday_counts, key=weekday_counts.get)
-                    history_lines.append(f"- 主な開催曜日: {most_common_weekday}")
-                
-                # 週の傾向（第何週か）
-                weeks = [self._get_week_of_month(e.date) for e in recent_events]
-                week_counts = {}
-                for w in weeks:
-                    week_counts[w] = week_counts.get(w, 0) + 1
-                
-                if week_counts:
-                    most_common_week = max(week_counts, key=week_counts.get)
-                    history_lines.append(f"- 主な開催週: 第{most_common_week}週")
-                
-                # 間隔の分析
-                if len(recent_events) >= 2:
-                    intervals = []
-                    for i in range(len(recent_events) - 1):
-                        interval = (recent_events[i].date - recent_events[i + 1].date).days
-                        intervals.append(interval)
-                    
-                    if intervals:
-                        avg_interval = sum(intervals) / len(intervals)
-                        history_lines.append(f"- 平均開催間隔: {avg_interval:.1f}日")
-                        
-                        # 隔週パターンの検出
-                        if 12 <= avg_interval <= 16:
-                            history_lines.append("- パターン: 隔週開催の可能性が高い")
-                        elif 27 <= avg_interval <= 32:
-                            history_lines.append("- パターン: 月1回開催の可能性が高い")
-            
-            return "\n".join(history_lines)
-            
-        except Exception:
-            logger.exception("Error getting recent events history")
-            return "過去の開催履歴: 取得エラー"
-    
-    def _generate_dates_by_llm(self, rule: RecurrenceRule, base_date: date, base_time: datetime.time, months: int, community: Optional["Community"] = None) -> List[date]:
-        """LLMを使用して日付リストを生成"""
-        if not rule.custom_rule:
-            return []
-        
-        end_date = base_date + timedelta(days=months * 30)
-        if rule.end_date and rule.end_date < end_date:
-            end_date = rule.end_date
-        
-        # 直近5回の開催履歴を取得
-        recent_events_history = self._get_recent_events_history(rule, base_date, community)
-        
-        prompt = f"""
-以下の条件で定期イベントの日付リストを生成してください。
-
-基準日: {base_date.strftime('%Y-%m-%d')} ({self._get_japanese_weekday(base_date.weekday())})
-基準時刻: {base_time.strftime('%H:%M')}
-生成期間: {base_date.strftime('%Y-%m-%d')} から {end_date.strftime('%Y-%m-%d')} まで
-定期ルール: {rule.custom_rule}
-
-{recent_events_history}
-
-出力形式:
-- YYYY-MM-DD形式の日付のJSONリスト
-- 基準日も含める
-- 日付は昇順でソート
-- 例: ["2024-01-01", "2024-01-15", "2024-02-01"]
-
-重要な注意事項:
-- 過去の開催履歴から主要なパターンを分析し、そのパターンを継続してください
-- 履歴にイレギュラーな開催（単発イベントや変則的な日程）が含まれる場合がありますが、それらは無視して主要なパターンのみを抽出してください
-- 隔週パターンで週がずれている場合は、最も新しい開催日から隔週で開催されるようにしてください
-- 「月1回」「月〇回」など月単位の周期の場合は、過去の開催曜日と週のパターンを最優先で考慮してください
-- 例：過去の履歴が主に第2金曜日のパターンなら、今後も第2金曜日を選んでください
-- イレギュラーな日程は生成せず、定期的なパターンのみを生成してください
-- 日本の祝日を考慮する場合は、2024年以降の実際の祝日を使用
-- 曜日は正確に計算すること
-- 存在しない日付（2月30日など）は含めない
-"""
-        
-        try:
-            dates = self._get_llm_service().generate_event_dates(prompt)
-            return sorted({generated_date for generated_date in dates if base_date <= generated_date <= end_date})
-        except Exception:
-            logger.exception("LLM date generation failed")
-
-        return []
-    
-    def preview_dates(self, frequency: str, custom_rule: str, base_date: date, base_time: datetime.time,
-                     interval: int = 1, week_of_month: Optional[int] = None, weekday: Optional[int] = None,
-                     months: int = 3, community: Optional["Community"] = None) -> Dict:
-        """プレビュー用に日付リストを生成"""
-        try:
-            # 一時的なRecurrenceRuleオブジェクトを作成（保存はしない）
-            rule = RecurrenceRule(
-                frequency=frequency,
-                interval=interval,
-                week_of_month=week_of_month,
-                custom_rule=custom_rule
-            )
-            
-            # MONTHLY_BY_WEEKの場合、weekdayからstart_dateを計算
-            if frequency == 'MONTHLY_BY_WEEK' and weekday is not None and week_of_month is not None:
-                # base_dateの月の第N曜日を計算してstart_dateとする
-                start_date = self._get_nth_weekday_of_month(
-                    base_date.replace(day=1),
-                    weekday,
-                    week_of_month
-                )
-                if start_date:
-                    rule.start_date = start_date
-            
-            dates = self.generate_dates(rule, base_date, base_time, months, community)
-            
-            return {
-                'success': True,
-                'dates': [d.strftime('%Y-%m-%d') for d in dates],
-                'count': len(dates)
-            }
-        except Exception as e:
-            return {
-                'success': False,
-                'error': str(e),
-                'dates': [],
-                'count': 0
-            }
-    
-    def create_recurring_events(self, community, rule: RecurrenceRule, base_date: date, 
-                              start_time: datetime.time, duration: int, months: int = 3) -> List[Event]:
-        """定期イベントのインスタンスを作成"""
-        dates = self.generate_dates(rule, base_date, start_time, months, community)
-        created_events = []
-        
-        # マスターイベントを作成（最初の日付）
-        if dates:
-            master_event = Event.objects.create(
-                community=community,
-                date=dates[0],
-                start_time=start_time,
-                duration=duration,
-                weekday=dates[0].strftime('%a').upper()[:3],
-                recurrence_rule=rule,
-                is_recurring_master=True
-            )
-            created_events.append(master_event)
-            
-            # 残りのインスタンスを作成
-            for event_date in dates[1:]:
-                # 既存のイベントがあるかチェック
-                existing = Event.objects.filter(
-                    community=community,
-                    date=event_date,
-                    start_time=start_time
-                ).first()
-                
-                if not existing:
-                    event = Event.objects.create(
-                        community=community,
-                        date=event_date,
-                        start_time=start_time,
-                        duration=duration,
-                        weekday=event_date.strftime('%a').upper()[:3],
-                        recurring_master=master_event
-                    )
-                    created_events.append(event)
-        
-        return created_events
+__all__ = ["RecurrenceService"]
